fusion serie/eth effectuée (comMgr), disparition (snif) de define_debug.h et de DEBUG remplacé par define_com_options.h, rename de SerialMgr et EthernetMgr en SerialInterface et EthernetInterface, creation d'AbstractComInterface, creation d'un typedef ( using pluôt ) pour les Headers, tout cela testé et fonctionel, reste plus qu'a adapter la deuxieme partie du protocole de com de william dessus, maintenant, bonne nuit!

+++ src/Sensors/SRF10.h
@@ -8,7 +8,7 @@
 class SRF10 : public AbstractI2CSensorUS
 {
 private:
-    uint8_t real_addr = BASE_OFFSET_ADDRESS;
+    uint16_t real_addr = BASE_OFFSET_ADDRESS;
     uint16_t last_distance_measured = 0xFFFF;
     bool waitingMeasure = false;
